Check if read IO object value is `blank?`
Reading from IO objects is implemented in the shared adapter, but the
verification for "blankness" is not applied for values read from these
objects, causing MultiJson to fallback to each adapter interpretation of
blank/invalid.
This results in different behaviour - depending on the adapter - for the
same input.

This commit solve this issue, making MultiJson always raise Parse
Error for blank?ish values read from IO objects.